Rename module CssCompiler to Assets
