+++ decidim-core/lib/decidim/assets.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Assets
+    autoload :Tailwind, "decidim/assets/tailwind"
+  end
+end
