+++ decidim-core/lib/decidim/assets/tailwind/tailwind.config.js.erb
