--- decidim-core/lib/decidim/css_compiler.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "decidim/css_compiler/tailwind"
-
-module Decidim
-  module CssCompiler
-  end
-end
